Reworked the jQuery to limit the number of taxis that can be selected to 1 at a time.
To make the functionallity different in the two tables, I had to add classes to them specifying the table that they belonged to; 'requestRow' and 'taxiRow'.
I thought to use id's but since they should be unique and we add the rows in a loop, i'd rather use them for the database ID's of the objects in the rows (to ease posting orders to the database).

Had to change from $(this).attr('class' == '...') to $(this).hasClass('...') for it to be able to work with multiple classes ('requestRow/taxiRow' and 'selectedRequest/selectedTaxi')..

The code is larger and more specific, but it works like a charm.